update render graph
