+++ osgraph-service-py/.gitignore
@@ -8,6 +8,7 @@ venv.bak/
 .env
 logs/
 *.log
+app/script/node_modules
 
 # Python cache
 __pycache__/
