+++ osgraph-service-py/app/script/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "script",
+  "version": "1.0.0",
+  "main": "g6_render.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "description": "",
+  "dependencies": {
+    "puppeteer": "^23.9.0"
+  }
+}
